requirements.txt
langchain
langgraph
openai
faiss-cpu
tiktoken
pymongo
pyyaml
